% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers.R
\docType{data}
\name{flowers}
\alias{flowers}
\title{flowers}
\format{A data frame with X rows and 9 variables}
\usage{
data(flowers)
}
\description{
A dataset containing flower counts per transect
}
\details{
\itemize{
  \item SiteID: Unique name of the sites where collected. See sites table for more info.
  \item Day: Day fruits were collected
  \item Month: Month fruits were collected
  \item Year: Year fruits were collected
  \item Round: identifies of each of the rounds done to a site
  \item Plot: Each of the 10 two by two meters plots where flowers where counted within a transect
  \item Plant.genus: Plant genus
  \item Plant.species: Plant species
  \item Flower.abundance: number of flowers counted per plot
  }
}
\keyword{datasets}
