% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_visits.R
\docType{data}
\name{single_visits}
\alias{single_visits}
\title{single visits}
\format{A data frame with X rows and Y variables}
\usage{
data(single_visits)
}
\description{
A dataset containing pollinator efficiency after a single visit to a virgin flower.
}
\details{
\itemize{
  \item SiteID: Unique name of the sites where collected. See `sites` table for more info.
  \item Name: Team collecting the data
  \item Day: Day single visit was recorded
  \item Month: Month single visit was recorded
  \item Year: Year single visit was recorded
  \item Time: Time single visit was recorded
  \item Weather: Weather (Sol = Sunny, Nubes = Cluody)
  \item Temperature: in Cº
  \item Humidity: in %
  \item Wind: (No = No wind; Suave = mild)
  \item Notes: Notes
  \item Visit.time.sec: Duration of the visit in seconds
  \item Stigma_id: Stigma ID for internal purposes
  \item Plant_genus: Plant genus
  \item Plant_species: Plant species
  \item Pollinator_id: Pollinator ID (vouchered at EBD)
  \item Pollinator_genus: Pollinator genus (including controls)
  \item Pollinator_species: Pollinator species. Need to be updated when  
  \item subspecie: when known.
  \item Pollinator_sex: male, female, worker
  \item morfospecie: As a note
  \item Det: Taxonomist
  \item pollen_tubes: number of pollen tubes developed after a single visit.
  }
}
\keyword{datasets}
