% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites.R
\docType{data}
\name{sites}
\alias{sites}
\title{Site information}
\format{A data frame with X rows and Y variables}
\usage{
data(sites)
}
\description{
A dataset containing site information
}
\details{
\itemize{
  \item SiteID: Unique name of the sites where collected. See sites table for more info.
  \item latitude: in decimal degrees
  \item longitude: in decimal degrees
  }
}
\keyword{datasets}
