% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits_pollinators_estimated.R
\docType{data}
\name{traits_pollinators_estimated}
\alias{traits_pollinators_estimated}
\title{Pollinator trait data}
\format{A data frame with X rows and Y variables}
\usage{
data(traits_pollinators_estimated)
}
\description{
A dataset containing pollinator trait data.
}
\details{
\itemize{
  \item Pollinator_gen_sp: Pollinator species name
  \item IT: Intertegular distance for bees. span between weings for flies, torax width for beetles. Wing span for buteflies. In mm
  \item family: taxonomic family
  \item order: taxonomic order
  \item tongue_length: tongue length in mm. Some estimated using Cariveau et al 2015
  \item pollinator: binary. Are considered pollintors or not.
  \item mouth_type: mouth type (chewing, lapping, sucking, biting, piercing)
  \item use_nectar_pollen: What adults use as resource.
  }
}
\references{
"Cariveau, Nayak, Bartomeus, Ascher, Winfree. 2015 The allometry of bee tongue length an its uses in ecology and evolution" http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0151482
}
\keyword{datasets}
